Fix structure of plist version
The conversion of the JSON version to Plist was not quite correct.
Some array tags should have been dict tags. This caused the file
not to load. For Sublime there is no issue because it preferentially
uses the YAML version. However the github/linquist project uses the
plist version for syntax highlighting of J code on Github. This has
meant that for the last month or so J code on Github would not have
been syntax highlighted. This commit should fix that once it is picked
up by the linguist project.